### 2019-07-26 #1598 @kyleniemeyer
Hi @chand-lab, thanks for your submission to JOSS. Please merge the PR I just submitted to your paper, which makes some necessary fixes to the authors and affiliation fields: https://github.com/chand-lab/openEyeTrack/pull/1

Also, please add ORCIDs for both authors; even after merging my PR, the paper will still not build correctly without those. (If you don't have ORCIDs, you can register for them here: https://orcid.org/register). Thanks!
