+++ paper/paper.md
@@ -11,7 +11,7 @@ authors:
     orcid: 
     affiliation: 1
   - name: Chandramouli Chandrasekaran
-    orcid: 
+    orcid: 0000-0002-1711-590X
     affiliation: "2, 3"
 affiliations:
  - name: Department of Biomedical Engineering, Boston University, Boston, MA 02215, USA
