### 2019-07-26 #1598 @chand-lab
Hi Kyle,

Yes. Adding orcids. I need to get my undergraduate to get one as well.

best,

Chand

On Fri, Jul 26, 2019 at 12:33 PM Kyle Niemeyer <notifications@github.com>
wrote:

> Hi @chand-lab <https://github.com/chand-lab>, thanks for your submission
> to JOSS. Please merge the PR I just submitted to your paper, which makes
> some necessary fixes to the authors and affiliation fields:
> chand-lab/openEyeTrack#1
> <https://github.com/chand-lab/openEyeTrack/pull/1>
>
> Also, please add ORCIDs for both authors; even after merging my PR, the
> paper will still not build correctly without those. (If you don't have
> ORCIDs, you can register for them here: https://orcid.org/register).
> Thanks!
>
