+++ README.md
@@ -21,7 +21,7 @@ We have had problems with 18.04 and recommend an install on the 16.04 system. We
  - Install git by typing: $ sudo apt-get install git-core
  - Clone repository by typing: $ git clone https://github.com/mailchand/openEyeTrack.git
  - After installing the GigE-V Framework, you can either manually set up Qt and OpenCV or run the installation script denoted as *softwareInstall* within the repository
-   - If you choose to run the installation script you may need to first convert it to an executable by typing: $ chmod +x software_install.sh
+   - If you choose to run the installation script you may need to first convert it to an executable by typing: $ chmod +x softwareInstall.sh
    - Occassionally, you may be prompted to press "y" to continue the installation process
    - Please explicitly run the last two commands at the bottom of this file concerning linking of libraries and make sure Ubuntu actually identifies where your libraries are.
 
