### 2019-07-26 342a6e1
Update paper.md
`tag` -> `tags`
