### 2019-07-26 #1598 @kyleniemeyer
@chand-lab my new PR should fix the issue: https://github.com/chand-lab/openEyeTrack/pull/2
