### 2019-07-26 3fcc579
Update paper.md
