### 2019-07-26 33fd1cf
Update paper.md
