+++ README.md
@@ -117,7 +117,7 @@ Use:
     sudo sh -c 'echo "/usr/local/lib" >> /etc/ld.so.conf.d/opencv.conf'
     sudo ldconfig
 
-For detailed usage examples, see [USAGE.md](docs/USAGE.md) found under the docs folder
+For detailed usage examples, see [USAGE.md](usage/USAGE.md) found under the docs folder
 
 
 
