### 2019-07-30 #1598 @cMadan
@chand-lab, can you clarify for me, what *specialised* hardware is needed for a reviewer to be able to evaluate this submission? How much of the submission can be evaluated without any specialised hardware?
