### 2019-07-30 #1598 @chand-lab
@cmadan, this is code for an open source eye tracker which can be used for research settings. It unfortunately needs a high speed camera from Teledyne DALSA and an infrared emitter for getting purkinje images of the eye. Teledyne DALSA has an open source linux API. Our contribution comes from developing a multithreaded C++ client that can leverage their API and then use opencv to detect the pupil and send this data to other downstream applications through UDP. I recognize now that the need for hardware may make it unsuitable for JOSS :).
