### 2019-07-30 #1598 @chand-lab
Reviewer suggestion: Jan Zimmerman (janz@umn.edu). Jan built the original open source non threaded video eye tracker. Our project builds on top of it.
