### 2019-07-31 #1598 @cMadan
@thejanzimmermann, great, I assume you have access to the requisite hardware? I'll try and also get someone who's previously reviewed for JOSS on-board as well to help review the code, but likely wouldn't have the requiste hardware to do a full review of the functionality.

@adswa, would you be available and interested in reviewing this submission?
