### 2019-08-02 #1598 @rougier
I'm currently involved in two reviews for JOSS (and I'm on vacation) so it might be better to ask someone else, sorry.
