### 2019-08-05 #1598 @mailchand
@cMadan  I would perhaps suggest @conradsnicta as a potential reviewer as well. He has C++ experience and also appears to be involved in computer vision and might actually be an ideal reviewer for this small package.
