### 2019-08-06 #1598 @conradsnicta
@cMadan @mailchand - I'll take a look, though it may take a little while (~2 weeks) as I'm juggling a few projects. I do not have the requisite hardware to do proper testing, so any review will necessarily not cover all the aspects of the software.
