### 2019-08-07 #1598 @cMadan
@conradsnicta, thank you for agreeing, that's fine with me. We do have another reviewer with access to the hardware, so just be as thorough as you can.
