### 2019-08-07 #1631 @whedon
[REVIEW]: openEyeTrack - A high speed multi-threaded eye tracker for head-fixed applications
**Submitting author:** @chand-lab (<a href="http://orcid.org/0000-0002-1711-590X">Chandramouli Chandrasekaran</a>)
**Repository:** <a href="https://github.com/chand-lab/openEyeTrack" target ="_blank">https://github.com/chand-lab/openEyeTrack</a>
**Version:** v1.0.0
