### 2019-08-08 #1631 @thejanzimmermann
The authors provide a logical and necessary extension of previously published work on high speed eye tracking. This is especially important since most human applications or other open source initiatives provide low tracking speeds not acceptable to research applications.

The software is easy to compile, very well documented and understandable and provides a great place to start for researchers.

One thing that should be noted is that the system does not handle ocular occlusions very well because it computes the pupil based on a circular fit rather than a center of mass of an arbitrarily shaped object. This is non problematic in healthy human subjects but can become problematic in animal models.

Overall a great piece of software, well documented.
