### 2019-08-08 #1631 @chand-lab
Dear @thejanzimmermann,

Thank you for the kind comments on the manuscript. As far as we understand it, the opencv blob detector which is what is used for detecting the pupil should function even with partially occluded pupils.

The circle is drawn as a visual aid to denote to the experimenter that a position was identified. We found that when windowing the image to only detect images within a particular region of interest, we could not find a reliable way to take the keypoints detected by the blob detector and  show what was detected. Instead, we used the center and the "size" of the blob to draw a red circle.

In tests with artificial stimuli, our blob detector is even able to detect small black crescents drawn on a piece of paper. Of course, if this issue persists in animal models, we will consider other approaches for detecting the pupil. We believe the multithreaded approach allows for a little bit more leeway in implementing more computationally expensive algorithms.

best,

Chand
