### 2019-08-08 #1631 @thejanzimmermann
Dear @chand-lab , I could not agree more with you ! and it's not a dealbreaker at all.

I can forward you experimental code on adjusted center of mass approaches that I have written that might at least mitigate this a little.

Best

Jan
