### 2019-08-16 #1631 @mailchand
Hi @thejanzimmermann,

Thanks! I think this would be a great addition to v2.0 of the Eye Tracker. It might also make sense to incorporate corneal reflections from the Purkinje image in future versions.

best,

Chand
