### 2019-09-10 cc3c207
Update paper.md
