+++ paper/paper.md
@@ -7,7 +7,7 @@ tags:
  - OpenCV
  - Teledyne DALSA
 authors:
-  - name: Jorge (Paolo) Casas
+  - name: Jorge Paolo Casas
     orcid: 0000-0002-9575-1717
     affiliation: 1
   - name: Chandramouli Chandrasekaran
@@ -31,7 +31,7 @@ When faced with a decision, an organism uses information gathered by its senses 
 
 ### Software and Hardware components 
 
-_openEyeTrack_ is a video based eye tracker that takes advantage of OpenCV [@opencv_library], a low cost high speed infrared camera and Gige-V APIs for Linux provided by Teledyne DALSA [@GigeV], the graphical user interface toolkit QT5 [@qt5] and the opencv based gui elements [@cvui], all of which can be downloaded for free. The only costs are from the hardware components such as the camera (Genie Nano M640 NIR, Teledyne DALSA, ~$450, ~730 frames per second) and infrared light source, an articulated arm to position the camera (Manfrotto: $130), a computer with one or more gigabit network interface cards, and a power over ethernet switch to power and receive data from the camera. 
+_openEyeTrack_ is a video based eye tracker that takes advantage of OpenCV [@opencv_library], a low cost high speed infrared camera and Gige-V APIs for Linux provided by Teledyne DALSA [@GigeV], the graphical user interface toolkit QT5 [@qt5] and the OpenCV based gui elements [@cvui], all of which can be downloaded for free. The only costs are from the hardware components such as the camera (Genie Nano M640 NIR, Teledyne DALSA, ~$450, ~730 frames per second) and infrared light source, an articulated arm to position the camera (Manfrotto: $130), a computer with one or more gigabit network interface cards, and a power over ethernet switch to power and receive data from the camera. 
 
 By using the Gige-V Framework to capture the frames from the DALSA camera and the OpenCV simple blob detector, _openEyeTrack_ is able to accurately calculate the position and area of the pupil. We include pupil size calculations because of its putative link to arousal levels and emotions of the observer [@doi:10.1111/j.1469-8986.2007.00606.x]. 
 
