+++ paper/paper.bib
@@ -13,7 +13,7 @@ year = {2008}
 }
 
 @online{openEyeTrack,
-  author = "Jorge (Paolo) Casas and Chandramouli Chandrasekaran",
+  author = "Jorge Paolo Casas and Chandramouli Chandrasekaran",
   title = "openEyeTrack - A high speed multi-threaded eye tracker for head-fixed applications",
   year = 2019,
   url = {https://github.com/chand-lab/openEyeTrack},
