### 2019-09-17 #1631 @conradsnicta
@cMadan - in progress; I'll add comments as I go along
