### 2019-09-17 #1631 @conradsnicta
@chand-lab @mailchand There seems to be a problem with the reference to OpenCV, which is listed as "Bradski, G. (2000). The OpenCV Library. Dr. Dobb’s Journal of Software Tools".  I can't seem to find the actual article.
* There is an apparent link listed at http://ftp.math.utah.edu/pub/tex/bib/dr-dobbs-2000.html#Bradski:2000:OL but following the link results in error 404.
* Searching https://www.drdobbs.com and http://www.ddj.com for OpenCV (directly and indirectly via Google) seems to yield nothing.
* Bradski's Google Scholar profile (https://scholar.google.com/citations?user=yeDFJgoAAAAJ&hl=en) has an entry for the article, but there is no link to the actual text / pdf.

So the cited article doesn't exist or has effectively ceased to exist. I suggest changing the reference to an article that is currently accessible.
