### 2019-09-17 #1631 @mailchand
Thanks @conradsnicta ! I was trying to find a way to cite openCV and chose what has been recommended online :). We used all the online resources when we were programming with openCV.  Instead, what we will do is to use this reference

[Open CV book](https://s3.amazonaws.com/academia.edu.documents/32167394/OReilly_Learning_OpenCV.pdf?response-content-disposition=inline%3B%20filename%3DLearning_OpenCV.pdf&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAIWOWYYGZ2Y53UL3A%2F20190917%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20190917T122132Z&X-Amz-Expires=3600&X-Amz-SignedHeaders=host&X-Amz-Signature=0b481a623ea494fa86d366dafd0c70c361d35c7352ca8caf73de3da498ad2302)
