+++ paper/paper.bib
@@ -28,6 +28,13 @@ year = {2008}
     year=2018,
     urldate={2019-07-26}
  }
+ 
+@book{bradski2008learning,
+  title={Learning OpenCV: Computer vision with the OpenCV library},
+  author={Bradski, Gary and Kaehler, Adrian},
+  year={2008},
+  publisher={" O'Reilly Media, Inc."}
+}
 
 @software{opencv_library,
     author = {Bradski, G.},
