### 2019-09-17 d00adcf
Update paper.md
