+++ paper/paper.bib
@@ -31,9 +31,12 @@ year = {2008}
  
 @book{bradski2008learning,
   title={Learning OpenCV: Computer vision with the OpenCV library},
-  author={Bradski, Gary and Kaehler, Adrian},
+  author={Gary Bradski and Adrian Kaehler},
   year={2008},
-  publisher={" O'Reilly Media, Inc."}
+  isbn = {9780596516130},
+  publisher={" O'Reilly Media, Inc."},
+  url={https://www.academia.edu/download/32167394/OReilly_Learning_OpenCV.pdf},
+  urldate={2019-09-17},
 }
 
 @software{opencv_library,
