### 2019-09-30 #1631 @conradsnicta
@chand-lab - The updated citation to Bradsky & Kaehler has several issues:

- The given URL (https://www.academia.edu/download/32167394/OReilly_Learning_OpenCV.pdf) ends up with Error 404.
- The publisher's name has unnecessary quotes around it.
- OpenCV is printed as lowercase opencv, twice.  In `paper.bib`, use curly braces around the word to force Bibtex to respect the upper case letters: {OpenCV}
