### 2019-09-30 #1631 @conradsnicta
@chand-lab - Problems with the references:

- The updated citation to Bradsky & Kaehler has several issues:
  - The given URL (https://www.academia.edu/download/32167394/OReilly_Learning_OpenCV.pdf) ends up with Error 404.
  - The publisher's name has unnecessary quotes around it.
  - OpenCV is printed as lowercase opencv, twice.  In `paper.bib`, use curly braces around the word to force Bibtex to respect the upper case letters: {OpenCV}

- Clicking on the DOI link for Zimmerann et al (2016) leads to an error. The DOI listed in `paper.bib` should only have the DOI number, not the URL.
