### 2019-09-30 #1631 @conradsnicta
@chand-lab - Grammar suggestions in the Algorithm section:
- It's not clear what is meant by "processed" in the sentence "as frames transition between the captured, processed and display processes". What do "captured process" and "processed process" mean ?  Perhaps the sentence would be clearer like so: "as frames transition between capture, processing and display stages". The "processing stage" should clarified or renamed, as it's too generic -- everything is "processing" in the end.
- Change "enable the different processes" to "enable the separate stages".
- Change "and stores in a queue" to "and stored in a queue".
