### 2019-09-30 #1631 @conradsnicta
@chand-lab
- In the Statement of Need section, there is a superfluous (or effectively circular) citation to Casas & Chandrasekaran (2019). This paper is about the very software being cited, so the citation is not necessary or needed.  You already have a link to GitHub page in the paragraph right after the list of limitations.
- Grammar suggestions in the Limitations section.  In the numbered list, there is no point in stating "first", "second", etc. The number already gives that information.
