### 2019-09-30 #1631 @mailchand
@conradsnicta - Thank you for a thorough read. I will attempt the address these issues and improve the manuscript ASAP. I think these are all excellent suggestions and will make the manuscript significantly better.
