+++ paper/paper.bib
@@ -51,6 +51,15 @@ year = {2008}
     year = {2000}
 }
 
+@book{Bradski:2008:LOE:1461412,
+ author = {Bradski, Dr. Gary Rost and Kaehler, Adrian},
+ title = {Learning Opencv, 1st Edition},
+ year = {2008},
+ isbn = {9780596516130},
+ edition = {First},
+ publisher = {O'Reilly Media, Inc.},
+} 
+
 @article{ZIMMERMANN2016138,
     title = "Oculomatic: High speed, reliable, and accurate open-source eye tracking for humans and non-human primates",
     journal = "Journal of Neuroscience Methods",
