### 2019-10-01 c31fcc1
Update paper.md
