+++ paper/paper.bib
@@ -53,7 +53,7 @@ year = {2008}
 
 @book{Bradski:2008:LOE:1461412,
  author = {Bradski, Dr. Gary Rost and Kaehler, Adrian},
- title = {Learning Opencv, 1st Edition},
+ title = {Learning OpenCV, 1st Edition},
  year = {2008},
  isbn = {9780596516130},
  edition = {First},
