+++ paper/paper.bib
@@ -52,10 +52,10 @@ year = {2008}
 }
 
 @book{Bradski:2008:LOE:1461412,
- author = {Bradski, Dr. Gary Rost and Kaehler, Adrian},
- title = {Learning OpenCV, 1st Edition},
- year = {2008},
- isbn = {9780596516130},
+ author = {Bradski, Gary and Kaehler, Adrian},
+ title = {Learning {OpenCV}, 1st Edition},
+ year = {2009},
+ isbn = {B005G9Z8I4},
  edition = {First},
  publisher = {O'Reilly Media, Inc.},
 } 
