### 2019-10-01 518a40c
Update paper.bib
