+++ paper/paper.bib
@@ -68,7 +68,6 @@ year = {2008}
     year = "2016",
     issn = "0165-0270",
     doi = "https://doi.org/10.1016/j.jneumeth.2016.06.016",
-    url = "http://www.sciencedirect.com/science/article/pii/S0165027016301467",
     author = "Jan Zimmermann and Yuriria Vazquez and Paul W. Glimcher and Bijan Pesaran and Kenway Louie",
     keywords = "Eye tracking, Search coil, Open-source, Non-human primate, Monkey, Rhesus"
 }
