+++ paper/paper.bib
@@ -56,9 +56,19 @@ year = {2008}
  title = {Learning {OpenCV}, 1st Edition},
  year = {2009},
  isbn = {B005G9Z8I4},
- edition = {First},
  publisher = {O'Reilly Media, Inc.},
 } 
+
+@article{opencv_library,
+    author = {Bradski, G.},
+    citeulike-article-id = {2236121},
+    journal = {Dr. Dobb's Journal of Software Tools},
+    keywords = {bibtex-import},
+    posted-at = {2008-01-15 19:21:54},
+    priority = {4},
+    title = {{The OpenCV Library}},
+    year = {2000}
+}
 
 @article{ZIMMERMANN2016138,
     title = "Oculomatic: High speed, reliable, and accurate open-source eye tracking for humans and non-human primates",
