### 2019-10-01 784b8fc
Update paper.md
