+++ paper/paper.bib
@@ -77,7 +77,7 @@ year = {2008}
     pages = "138 - 146",
     year = "2016",
     issn = "0165-0270",
-    doi = {https://doi.org/10.1016/j.jneumeth.2016.06.016},
+    doi = "https://doi.org/10.1016/j.jneumeth.2016.06.016",
     author = "Jan Zimmermann and Yuriria Vazquez and Paul W. Glimcher and Bijan Pesaran and Kenway Louie",
     keywords = "Eye tracking, Search coil, Open-source, Non-human primate, Monkey, Rhesus"
 }
