### 2019-10-01 247f7a8
Update paper.bib
