+++ paper/paper.bib
@@ -30,13 +30,11 @@ year = {2008}
  }
  
 @book{bradski2008learning,
-  title={Learning OpenCV: Computer vision with the OpenCV library},
+  title={Learning OpenCV 3: Computer vision in C++ with the OpenCV library},
   author={Gary Bradski and Adrian Kaehler},
-  year={2008},
-  isbn = {9780596516130},
+  year={2016},
+  isbn = {978-1491937990},
   publisher={" O'Reilly Media, Inc."},
-  url={https://www.academia.edu/download/32167394/OReilly_Learning_OpenCV.pdf},
-  urldate={2019-09-17},
 }
 
 @software{opencv_library,
@@ -84,7 +82,7 @@ year = {2008}
 
 @software{GigeV,
     author = {{Teledyne DALSA}},
-    title = {Gige-V Framework for Linux},
+    title = {{GigE-V} Framework for Linux},
     url = {https://www.teledynedalsa.com/en/support/downloads-center/software-development-kits/132/},
     version = {2.10},
     date = {2018-09-14}
