+++ paper/paper.bib
@@ -28,8 +28,17 @@ year = {2008}
     year=2018,
     urldate={2019-07-26}
  }
- 
-@book{bradski2008learning,
+
+@online{openCVcomputerlibrary,
+  title={OpenCV: Open Source Computer Vision Library},
+  author={OpenCV},
+  year={2015},
+  url = {https://github.com/opencv/opencv},
+  urldate={2019-10-01}
+}
+
+
+@book{bradski2016learning,
   title={Learning OpenCV 3: Computer vision in C++ with the OpenCV library},
   author={Gary Bradski and Adrian Kaehler},
   year={2016},
