+++ paper/paper.bib
@@ -43,7 +43,7 @@ year = {2008}
   author={Gary Bradski and Adrian Kaehler},
   year={2016},
   isbn = {978-1491937990},
-  publisher={" O'Reilly Media, Inc."},
+  publisher={O'Reilly Media, Inc.},
 }
 
 @software{opencv_library,
