+++ paper/paper.bib
@@ -12,6 +12,15 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1469-8986.2007.00606
 year = {2008}
 }
 
+@manual{opencvmanual,
+  title = {The OpenCV Reference Manual},
+  organization = {OpenCV},
+  edition = {4.1.1},
+  month = {July},
+  year = {2019},
+  notes = {See \url{https://docs.opencv.org/4.1.1/}}
+  }
+
 @online{openEyeTrack,
   author = "Jorge Paolo Casas and Chandramouli Chandrasekaran",
   title = "openEyeTrack - A high speed multi-threaded eye tracker for head-fixed applications",
