+++ paper/paper.bib
@@ -12,13 +12,11 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1469-8986.2007.00606
 year = {2008}
 }
 
-@manual{opencvmanual,
-  title = {The OpenCV Reference Manual},
-  organization = {OpenCV},
-  edition = {4.1.1},
-  month = {July},
-  year = {2019},
-  notes = {See \url{https://docs.opencv.org/4.1.1/}}
+@online{opencvmanual,
+  title = {The OpenCV Reference Manual, Edition 4.1.1},
+  author = {OpenCV},
+  url = {https://docs.opencv.org/4.1.1/},
+  urldate = {2019-07-01}
   }
 
 @online{openEyeTrack,
