+++ paper/paper.bib
@@ -39,9 +39,9 @@ year = {2008}
 @online{openCVcomputerlibrary,
   title={OpenCV: Open Source Computer Vision Library},
   author={OpenCV},
-  year={2015},
+  year={2019},
   url = {https://github.com/opencv/opencv},
-  urldate={2019-10-01}
+  urldate={2019-07-01}
 }
 
 
