### 2019-10-01 9a085f1
Update paper.md
