+++ paper/paper.bib
@@ -15,6 +15,7 @@ year = {2008}
 @online{opencvmanual,
   title = {The OpenCV Reference Manual, Edition 4.1.1},
   author = {OpenCV},
+  year = {2019},
   url = {https://docs.opencv.org/4.1.1/},
   urldate = {2019-07-01}
   }
