### 2019-10-01 d4a069c
Update paper.bib
