+++ paper/paper.bib
@@ -13,7 +13,7 @@ year = {2008}
 }
 
 @online{opencvmanual,
-  title = {The OpenCV Reference Manual, Edition 4.1.1},
+  title = {The {OpenCV} Reference Manual, Edition 4.1.1},
   author = {OpenCV},
   year = {2019},
   url = {https://docs.opencv.org/4.1.1/},
@@ -31,14 +31,14 @@ year = {2008}
 
 @online{cvui,
     author = "Fernando Bevilacqua",
-    title = "A (very) simple UI lib built on top of OpenCV drawing primitives",
+    title = "A (very) simple {UI} lib built on top of {OpenCV} drawing primitives",
     url={https://github.com/Dovyski/cvui},
     year=2018,
     urldate={2019-07-26}
  }
 
 @online{openCVcomputerlibrary,
-  title={OpenCV: Open Source Computer Vision Library},
+  title={OpenCV: {Open} {Source} {Computer} {Vision} {Library}},
   author={OpenCV},
   year={2019},
   url = {https://github.com/opencv/opencv},
