### 2019-10-01 8e1f923
Update paper.md
