### 2019-10-01 4dfa540
Update paper.md
