+++ paper/paper.bib
@@ -47,7 +47,7 @@ year = {2008}
 
 
 @book{bradski2016learning,
-  title={Learning OpenCV 3: Computer vision in C++ with the OpenCV library},
+  title={Learning {OpenCV} 3: Computer vision in C++ with the {OpenCV} library},
   author={Gary Bradski and Adrian Kaehler},
   year={2016},
   isbn = {978-1491937990},
