### 2019-10-01 8b35726
Update paper.md
