+++ paper/paper.md
@@ -46,12 +46,12 @@ _openEyeTrack_ is modeled on other open-source eye trackers currently available 
 
 ### Algorithm
 
-Figure 1 shows the workflow for _openEyeTrack_. As frames transition between the capture, detection, and display stages, they are stored in queues which enable the separate stages to run independently and allow for asynchronous capture, detection, and display. Once the camera grabs a new frame, it is briefly stored in the Genicam memory buffers before being extracted and packaged by the "capture thread" into a struct and stored in a queue. This approach allows for the sequence of acquisition frames to be preserved and the frame acquisition process to take place without being slowed down by the blob detection threads or the display thread. The frames in the "capture queue" are then popped off by the "n" (user-specified) detection thread(s). Each "detection thread" takes the data from the "capture queue," converts it into an OpenCV Mat object, applies the OpenCV blob detection algorithm, notes the key features. Each detection thread also outputs the blob position as text on the frame and draws a circle around the blob. All of these steps are very time consuming, which is why we recommend initializing multiple threads for higher performance. Once the detection stage has completed, the frames are stored in a "display queue" that the "display thread" will grab from to show the images. The detection threads also package the frame and keypoints information into a struct object and stores it in a "network queue." The "network thread" reads from this queue and sends out packets over a UDP socket for downstream applications.
+Figure 1 shows the workflow for _openEyeTrack_. As frames transition between the capture, detection, and display stages, they are stored in queues which enable the separate stages to run independently and allow for asynchronous capture, detection, and display. Once the camera grabs a new frame, it is briefly stored in the Genicam memory buffers before being extracted and packaged by the "capture thread" into a struct and stored in a queue. This approach allows for the sequence of acquisition frames to be preserved and the frame acquisition process to take place without being slowed down by the blob detection threads or the display thread. The frames in the "capture queue" are then popped off by the "n" (user-specified) detection thread(s). Each "detection thread" takes the data from the "capture queue," converts it into an OpenCV Mat object, applies the OpenCV blob detection algorithm, and notes the key features. Each detection thread also outputs the blob position as text on the frame and draws a circle around the blob. All of these steps are very time consuming, which is why we recommend initializing multiple threads for higher performance. Once the detection stage has completed, the frames are stored in a "display queue" that the "display thread" will grab from to show the images. The detection threads also package the frame and keypoints information into a struct object and stores it in a "network queue." The "network thread" reads from this queue and sends out packets over a UDP socket for downstream applications such as a task control computer.
 
  
 ### Performance 
 
-Under the conditions at the time of development, we were able to achieve frame acquisition frame rates of up to 715 fps and display rates of up to 145 fps. Although more threads, in theory, should increase speed, four processing threads were sufficient to keep up with the camera. Performance improved when we used the _gev_nettweak_ tool provided by Teledyne Dalsa, which adjusts various features for the network buffers allowing higher throughput transmission from the camera to the computer. Additionally, environmental lighting significantly affects the speed at which the blob detection occurs. The OpenCV blob detector by default looks for black blobs, and thus more light allows for easier detection by increasing the contrast between darker and lighter areas of the image. To facilitate blob detection, we also apply binary thresholding to the images. The user can also specify a region of interest for the blob detector, which again improves processing time. For eye tracking, it is necessary to have an infrared IR light source to obtain a Purkinje image with increased contrast between the pupil and the surrounding regions.
[…]
