### 2019-10-01 c2847a7
Update paper.md
