### 2019-10-01 d4f599f
Update paper.md
