+++ paper/paper.md
@@ -41,7 +41,7 @@ By using the GigE-V Framework to capture the frames from the DALSA camera and th
 
 ### Multithreading provides improvements over existing open source solutions
 
-_openEyeTrack_ is modeled on other open-source eye trackers currently available such as _“Oculomatic”_ [@ZIMMERMANN2016138]. However, most of these programs are single-threaded: the frames are captured, processed, and displayed sequentially, only executing the next stage once the previous stage completes its processing. Although single-threaded methods have become more effective over the years, these stages are time-consuming and can limit the overall performance. In order to increase performance, _openEyeTrack_ was developed as a multithreaded application. The capture, display, data transmission, and most importantly, processing components all happen within their separate threads. By incorporating multiple threads, the processing speed of the frames can match the frame capture rate of the camera, allowing for the lossless processing of data.
+_openEyeTrack_ is modeled on other open-source eye trackers currently available such as _“Oculomatic”_ [@ZIMMERMANN2016138]. However, most of these programs are single-threaded: the frames are captured, analyzed, and displayed sequentially, only executing the next stage once the previous stage completes its processing. Although single-threaded methods have become more effective over the years, these stages are time-consuming and can limit the overall performance. In order to increase performance, _openEyeTrack_ was developed as a multithreaded application. The capture, display, data transmission, and most importantly, pupil detection components all happen within their separate threads. By incorporating multiple threads, the processing speed of the frames can match the frame capture rate of the camera, allowing for the lossless processing of data.
 
 
 ### Algorithm
