### 2019-10-02 #1631 @mailchand
Hello @conradsnicta and @cMadan

We have now revised the paper. I think we have caught all the issues that @conradsnicta raised and have resolved them. We apologize if we have missed something because the issue list is mixed in with all the generate pdf commands.

- All links have been checked.
- Citations to OpenCV now cite the websites and the book from Bradski and Kaehler minus the URL.
- Poor grammar has been corrected.
- Statement of need updated to reflect ongoing experiments rather than dangling at the end of the paper.

I believe these are the list of identified issues. Attached is a PDF with responses as well. The paper.md file has been updated and the pdf generated.

best,

Chand
[Responses.pdf](https://github.com/openjournals/joss-reviews/files/3682350/Responses.pdf)
