### 2019-10-04 #1631 @conradsnicta
@cMadan - As I don't have the necessary hardware (specific camera), I am unable to tick the "Functionality" and "Performance" boxes within the review. Other than that it looks mostly okay.

There is a limitation/problem which must be pointed out. In README.md, it's stated that the software specs include "Ubuntu 16.04". On an Ubuntu 18.04 system, after manually installing qt5 development packages (which necessitated downgrades of several other packages), I did manage to run the supplied "softwareInstall.sh" script. Without the manual intervention the installation script appeared to break.

At this point Ubuntu 16.04 is pretty old.  According to the Ubuntu [release cycle](https://ubuntu.com/about/release-cycle), 16.04 ceased to obtain hardware updates in late 2018, meaning that it's not guaranteed not install on new machines. Maintenance updates (eg. security) for 16.04 will cease in early 2021.

Given that there is already apparent installation issues on Ubuntu 18.04 and that 16.04 is not really a viable option for new machines, the software presented in this paper is subject to bit rot. In other words, it's longevity is at risk without continual maintenance.  There is material risk that this software will not install as is on Ubuntu 20.04.
