### 2019-10-07 #1631 @mailchand
@conradsnicta - Great re: Functionality and Performance.

Re: 16.04 vs. 18.04 -- Thanks for raising this issue. We have an 18.04 test suite as well and now re test the software on 18.04 and see if I can mitigate the issue you have raised. We will attempt to  provide a revised install script for 18.04. Your point about ensuring 18.04 and 20.04 compatibility is well taken. Here is our plan:

1) Attempt to fix softwareInstall.sh to ensure 18.04 compatibility
2) If this fails miserably, note the limitation/problem associated with it and our plans for continued development to ensure forward compatibility.

best,

Chand
