+++ installDepsForDALSA.sh
@@ -1,0 +1,2 @@
+sudo apt-get install gcc g++ make libx11-dev libxext-dev libgtk-3-dev libglade2-0 libglade2-dev
+sudo apt-get install libpcap0.8 libcap2 ethtool
