### 2019-10-07 #1631 @mailchand
Hi @conradsnicta and @cMadan

In response to the concern @conradsnicta raised. I followed the steps outlined in Readme.md to install openEyeTrack on an 18.04.3 ubuntu install and a 19.04 install. Both of these were done on a Zotac EN51050 machine without any existing software in them and I did do clean installs. My approach is typical of experimental rigs where the computers are often dedicated to certain functions. I for instance do this with my neurophysiology hardware.

Under such conditions, the software cleanly installs on both Ubuntu versions and also did not need manual intervention beyond the steps previously documented in Readme.md (installing Teledyne DALSA's GigE-V API etc and its attendant dependencies). We were able to get this up and running in half an hour even when starting from a completely vanilla 18.04 or 19.04 install.
To help potential users more, the DALSA dependencies are now also provided as a shell script. Of course, we will never be able to support unique development environments and we hope these users will raise issues on our github page.

As evidence, I am attaching screenshots of openEyeTrack running on both 18.04 and 19.04. So we are very confident that openEyeTrack will be compatible with future hardware and also with newer versions of Ubuntu and not be susceptible to the bit rot issues. For Ubuntu 18.04 and 19.04 The Qt versions are also different and so openEyeTrack is not dependent on a particular version of Qt. We emphasize that for both we just used the qt5-default install line in the script which seems to work for both Ubuntu versions. We cannot speak to Ubuntu 20.04 because it is non existent at the moment.

@conradsnicta: If you are willing to provide what steps you used when the install broke for you, we would like to include it in the Readme.md  to say what to do if the install breaks on the qt install line. We can add that in as a warning of potential issues that could arise to help users of this software. I have now updated the Readme.md to say that you can use it on both 18.04 and 19.04.

![Screenshot from 2019-10-07 18-23-12](https://user-images.githubusercontent.com/1933550/66353462-14500000-e930-11e9-89c2-6485c657d634.png)
![Screenshot from 2019-10-07 16-12-59](https://user-images.githubusercontent.com/1933550/66353463-14500000-e930-11e9-8813-e7023037d7fc.png)

Please let us know if there are any other issues to resolve!

best,

Chand
