### 2019-10-08 #1631 @conradsnicta
@mailchand @cMadan I have a separate box with an Ubuntu 18.04 installation, which has proprietary AMD and NVIDIA drivers and repos enabled (for GPU related work via OpenCL and CUDA). Other than the installation wasn't modified to my knowledge.

It's entirely possible that the proprietary drivers and/or extra repos may hinder with the installation of openEyeTrack. This in turn may indicate that a pure vanilla installation of 18.04 may work without flaws openEyeTrack.

I can try to install openEyeTrack within a virtual machine installation of 18.04, though this may take a few days.
