### 2019-10-08 #1631 @mailchand
Hi @conradsnicta,

Hmm, that is entirely possible. I am sure you have way more experience dealing with Linux issues given your experience with intensive development in Linux. I personally encounter issues when trying to install new software because my dependencies are not properly structured. For instance, when trying to install DeepLabCut, an open source behavioral tracking tool that uses Tensorflow and CUDA, there have been times where I struggled to get it working.

I also ran the shell script as is. It contains sudo apt-get update and sudo apt-get upgrade commands which may or may not be helpful.

It is up to you and and @cMadan to make the decision to install on a ubuntu 18.04 virtual machine for additional tests because of your role a reviewer and @cMadan being the editor. I am also willing to add a caveat in the Readme.md if it is more expedient to say that we have only tested on clean installs and that we recommend such a strategy.

Thanks for all the care and attention to this paper.

regards,

Chand
