### 2019-10-17 #1631 @conradsnicta
@cMadan - I've confirmed that `softwareInstall.sh` works on a fresh installation of Ubuntu 18.04.3.  I used a virtual machine with 4 GB ram and 32 GB disk space.  (Using lower amounts, 2 GB ram and 16 GB disk space, led to "internal compiler error").

@mailchand - I believe the review is done from my side. I ticked all the boxes apart from "Functionality" and "Performance". As per https://github.com/openjournals/joss-reviews/issues/1631#issuecomment-538315331 I do not have the necessary hardware to confirm the details in these 2 boxes. However, @thejanzimmermann has ticked them.
