### 2019-10-17 #1631 @chand-lab
@conradsnicta: Thank you so much for all the diligent work to review this paper. Paolo and I appreciate it as we encourage people to use it widely.

@cMadan: Could you let us know what next steps we need to do to wrap up the paper and if there are any remaining issues on your end?

Thanks for a fun review experience!
