### 2019-10-18 #1631 @cMadan
@chand-lab, it looks like two of your references didn't resolve correctly in the PDF. See the second line of the "Software and Hardware components" section.
