+++ paper/paper.md
@@ -35,7 +35,7 @@ When faced with a decision, an organism uses information gathered by its senses 
 
 ### Software and Hardware components 
 
-_openEyeTrack_ is a video-based eye-tracker that takes advantage of OpenCV [@openCVcomputerlibrary, @opencvmanual, @bradski2016learning], a low-cost, high-speed infrared camera and GigE-V APIs for Linux provided by Teledyne DALSA [@GigeV], the graphical user interface toolkit QT5 [@qt5] and cvui, the OpenCV based GUI [@cvui]. All of the software components are freely available. The only costs are from the hardware components such as the camera ([Genie Nano M640 NIR](https://www.teledynedalsa.com/en/products/imaging/cameras/genie-nano-1gige/), Teledyne DALSA, ~$450, ~730 frames per second) and infrared light source, an articulated arm to position the camera (Manfrotto: $130), a computer with one or more gigabit network interface cards, and a power over ethernet switch to power and receive data from the camera. 
+_openEyeTrack_ is a video-based eye-tracker that takes advantage of OpenCV [@openCVcomputerlibrary,@opencvmanual,@bradski2016learning], a low-cost, high-speed infrared camera and GigE-V APIs for Linux provided by Teledyne DALSA [@GigeV], the graphical user interface toolkit QT5 [@qt5] and cvui, the OpenCV based GUI [@cvui]. All of the software components are freely available. The only costs are from the hardware components such as the camera ([Genie Nano M640 NIR](https://www.teledynedalsa.com/en/products/imaging/cameras/genie-nano-1gige/), Teledyne DALSA, ~$450, ~730 frames per second) and infrared light source, an articulated arm to position the camera (Manfrotto: $130), a computer with one or more gigabit network interface cards, and a power over ethernet switch to power and receive data from the camera. 
 
 By using the GigE-V Framework to capture the frames from the DALSA camera and the OpenCV simple blob detector, _openEyeTrack_ can accurately estimate the position and area of the pupil. We include pupil size calculations because of its putative link to arousal levels and emotions of the subject [@doi:10.1111/j.1469-8986.2007.00606.x]. 
 
