### 2019-10-18 #1631 @chand-lab
> @chand-lab, it looks like two of your references didn't resolve correctly in the PDF. See the second line of the "Software and Hardware components" section.

This has been fixed. I think the issue was you need to separate by ; for multiple references instead of a ,. Apologies for the error!
