### 2019-10-22 #1631 @chand-lab
Hi @cMadan,

Terrific. Please see mixed in below. I believe all the steps  are done!

> @chand-lab, everything looks good to me!
>
> To move forward with accepting your submission, there are a few last things to take care of:
>
> * [X]  Make a tagged release of your software, and list the version tag of the archived version here.

v1.0.0 - Available here
https://github.com/chand-lab/openEyeTrack/tree/v1.0.0

> * [X]  Archive the reviewed software in Zenodo

Done!

> * [ ]  Check the Zenodo deposit has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.
 > * [X] Authors
 > * [X] Paper title is correct
 > * [X] Added ORCID

> * [X]  List the Zenodo DOI of the archived version here.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3515534.svg)](https://doi.org/10.5281/zenodo.3515534)

>
> You may find this helpful: https://guides.github.com/activities/citable-code/

Please let me know if I missed something :).

Warm Regards,

Chand
