### 2019-10-23 #1631 @kthyng
Congratulations @chand-lab on your JOSS submission!! 🎉 🎉

Thank you to @conradsnicta and @thejanzimmermann for your time and expertise reviewing this submission. Thanks also to @cMadan for editing.
